requirements.txt
click==8.1.8
guidellm @ git+https://github.com/neuralmagic/guidellm.git@55c65c4
modal==0.73.116
numpy==2.2.4
PyYAML==6.0.2
SQLAlchemy==2.0.39